Separate out buildGetCertificateFunc for go1.9
